Make these tests 64-bit clean.
git-svn-id: svn://svn.valgrind.org/valgrind/branches/PTRCHECK@8496 a5019735-40e9-0310-863c-91ae7b9d1cf9

+++ exp-ptrcheck/tests/cmp.c
@@ -9,7 +9,7 @@ int main(void)
    char* buf = malloc(sizeof(char) * 6);
 
    // Known zero non-pointer
-   char* nz = (char*)((int)buf^(int)buf);  // known non-pointer
+   char* nz = (char*)((long)buf^(long)buf);  // known non-pointer
 
    // Unknown zero nonptr;  make them zero but unknown
    char* unz;
